Moved CTagsRunner command into builder package
