+++ src/arduino.cc/builder/test/ctags_runner_test.go
@@ -30,15 +30,15 @@
 package test
 
 import (
-	"arduino.cc/builder"
-	"arduino.cc/builder/constants"
-	"arduino.cc/builder/ctags"
-	"arduino.cc/builder/types"
-	"github.com/stretchr/testify/require"
 	"os"
 	"path/filepath"
 	"strings"
 	"testing"
+
+	"arduino.cc/builder"
+	"arduino.cc/builder/constants"
+	"arduino.cc/builder/types"
+	"github.com/stretchr/testify/require"
 )
 
 func TestCTagsRunner(t *testing.T) {
@@ -71,7 +71,7 @@ func TestCTagsRunner(t *testing.T) {
 		&builder.PrintUsedLibrariesIfVerbose{},
 		&builder.WarnAboutArchIncompatibleLibraries{},
 		&builder.CTagsTargetFileSaver{Source: &ctx.Source, TargetFileName: constants.FILE_CTAGS_TARGET},
-		&ctags.CTagsRunner{},
+		&builder.CTagsRunner{},
 	}
 
 	for _, command := range commands {
@@ -121,7 +121,7 @@ func TestCTagsRunnerSketchWithClass(t *testing.T) {
 		&builder.PrintUsedLibrariesIfVerbose{},
 		&builder.WarnAboutArchIncompatibleLibraries{},
 		&builder.CTagsTargetFileSaver{Source: &ctx.Source, TargetFileName: constants.FILE_CTAGS_TARGET},
-		&ctags.CTagsRunner{},
+		&builder.CTagsRunner{},
 	}
 
 	for _, command := range commands {
@@ -169,7 +169,7 @@ func TestCTagsRunnerSketchWithTypename(t *testing.T) {
 		&builder.PrintUsedLibrariesIfVerbose{},
 		&builder.WarnAboutArchIncompatibleLibraries{},
 		&builder.CTagsTargetFileSaver{Source: &ctx.Source, TargetFileName: constants.FILE_CTAGS_TARGET},
-		&ctags.CTagsRunner{},
+		&builder.CTagsRunner{},
 	}
 
 	for _, command := range commands {
@@ -216,7 +216,7 @@ func TestCTagsRunnerSketchWithNamespace(t *testing.T) {
 		&builder.PrintUsedLibrariesIfVerbose{},
 		&builder.WarnAboutArchIncompatibleLibraries{},
 		&builder.CTagsTargetFileSaver{Source: &ctx.Source, TargetFileName: constants.FILE_CTAGS_TARGET},
-		&ctags.CTagsRunner{},
+		&builder.CTagsRunner{},
 	}
 
 	for _, command := range commands {
@@ -262,7 +262,7 @@ func TestCTagsRunnerSketchWithTemplates(t *testing.T) {
 		&builder.PrintUsedLibrariesIfVerbose{},
 		&builder.WarnAboutArchIncompatibleLibraries{},
 		&builder.CTagsTargetFileSaver{Source: &ctx.Source, TargetFileName: constants.FILE_CTAGS_TARGET},
-		&ctags.CTagsRunner{},
+		&builder.CTagsRunner{},
 	}
 
 	for _, command := range commands {
